+++ src/index.ts
@@ -2,3 +2,4 @@ export * from "./types.js";
 export * from "./client.js";
 export * from "./errors.js";
 export * from "./webhooks.js";
+export * from "./protect.js";